Rewrite functionality to save changes based on textarea's value
The functionality before was much too complicated
in the sense that it required rewrite many keys'
natural behavior (ex. the user experience usually
allows one to hold the 'Backspace' key and delete
a section of text for as long as the key is pressed
down; the functionality before only registered backspaces
only ONCE every time the key was PRESSED). Relying
on the dynamically changing values of the textarea
to set the 'text' variable is much simpler in this
regard and for saving/retrieving changes upon refreshes.