Create save document button
The save document button must receive props from
the document component in order to store the document's
contents and title inside the 'documents' collection
that sits in the database. Because of this, the
document component and all its methods have been
rewritten as a child of the home component for
convenience.